Split monolithic app.py into server/ package
Pure refactoring -- no behavior changes. All 262 passing tests remain
green (30 pre-existing auth test failures unchanged).

Structure:
  server/__init__.py         - Flask app factory (create_app)
  server/security.py         - CSRF, session token auth, security headers
  server/db.py               - SQLite init, connection helpers, schema
  server/routes/library.py   - Library config, DICOM scanning, folder source
  server/routes/test_data.py - Test mode endpoints (/api/test-data/*)
  server/routes/study_notes.py - Study/series description endpoints
  server/routes/comments.py  - Comment CRUD endpoints
  server/routes/reports.py   - Report upload/download/delete, migration

app.py is now a thin entry point: imports create_app(), exposes the
Flask `app` instance for `flask run`, and keeps the __main__ block
for `python app.py` startup.
