+++ server/routes/__init__.py
@@ -1,0 +1,5 @@
+"""
+Flask route blueprints for the DICOM Viewer server.
+
+Copyright (c) 2026 Divergent Health Technologies
+"""
